♻️ Refactor out some builder and designer methods
