+++ bluemira/base/builder.py
@@ -26,12 +26,12 @@ Interfaces for builder classes.
 from __future__ import annotations
 
 import abc
-from typing import Dict, List, Optional, Type, Union
+from typing import Dict, List, Optional, Union
 
 from bluemira.base.components import Component
 from bluemira.base.designer import Designer
-from bluemira.base.error import BuilderError
 from bluemira.base.parameter_frame import NewParameterFrame as ParameterFrame
+from bluemira.base.parameter_frame import ParameterSetup
 from bluemira.utilities.plot_tools import set_component_view
 
 BuildConfig = Dict[str, Union[int, float, str, "BuildConfig"]]
@@ -82,7 +82,7 @@ class ComponentManager(abc.ABC):
         return self._component
 
 
-class Builder(abc.ABC):
+class Builder(ParameterSetup):
     """
     Base class for component builders.
 
@@ -99,6 +99,7 @@ class Builder(abc.ABC):
     -----
     If there are no parameters associated with a concrete builder, set
     `param_cls` to `None` and pass `None` into this class's constructor.
+    If param_cls is not `None` `param_cls` is set up with an empty dictionary.
     """
 
     def __init__(
@@ -107,18 +108,12 @@ class Builder(abc.ABC):
         build_config: Dict,
         designer: Optional[Designer] = None,
     ):
-        super().__init__()
+        super().__init__(params)
         self.name = build_config.get(
             "name", _remove_suffix(self.__class__.__name__, "Builder")
         )
-        self.params = self._init_params(params)
         self.build_config = build_config
         self.designer = designer
-
-    @abc.abstractproperty
-    def param_cls(self) -> Union[Type[ParameterFrame], None]:
-        """The class to hold this builder's parameters."""
-        pass
 
     @abc.abstractmethod
     def build(self) -> ComponentManager:
@@ -154,20 +149,3 @@ class Builder(abc.ABC):
         set_component_view(component.get_component("xy"), "xy")
 
         return component
-
-    def _init_params(
-        self, params: Union[Dict, ParameterFrame, None]
-    ) -> Union[ParameterFrame, None]:
-        if self.param_cls is None:
-            if params is None:
-                # Case for where there are no parameters associated with this builder
-                return params
-            else:
-                raise BuilderError("Cannot process parameters, 'param_cls' is None.")
-        elif isinstance(params, dict):
-            return self.param_cls.from_dict(params)
-        elif isinstance(params, ParameterFrame):
-            return params
-        raise TypeError(
-            f"Cannot interpret type '{type(params)}' as {self.param_cls.__name__}."
-        )
