+++ bluemira/base/parameter_frame/__init__.py
@@ -21,7 +21,7 @@
 """Module containing classes related to NewParameter and NewParameterFrame"""
 
 from bluemira.base.parameter_frame._decorator import parameter_frame
-from bluemira.base.parameter_frame._frame import NewParameterFrame
+from bluemira.base.parameter_frame._frame import NewParameterFrame, ParameterSetup
 from bluemira.base.parameter_frame._parameter import NewParameter
 
-__all__ = ["NewParameter", "NewParameterFrame", "parameter_frame"]
+__all__ = ["NewParameter", "NewParameterFrame", "ParameterSetup", "parameter_frame"]
